+++ cyphon/receiver/__init__.py
@@ -14,34 +14,3 @@
 #
 # You should have received a copy of the GNU General Public License
 # along with Cyphon Engine. If not, see <http://www.gnu.org/licenses/>.
-"""
-
-"""
-
-# third party
-from django.conf import settings as conf
-from django.db import models
-
-# local
-from sifter.chutes.models import Chute, ChuteManager
-from sifter.logsifter.logsieves.models import LogSieve
-from sifter.logsifter.logmungers.models import LogMunger
-
-
-class LogChuteManager(ChuteManager):
-    """
-    Adds methods to the default model manager.
-    """
-
-    settings = conf.LOGSIFTER
-
-
-class LogChute(Chute):
-    """
-
-    """
-    sieve = models.ForeignKey(LogSieve, null=True, blank=True, default=None,
-                              related_name='chutes', related_query_name='chute')
-    munger = models.ForeignKey(LogMunger)
-
-    objects = LogChuteManager()