+++ cyphon/receiver/receiver.py
@@ -27,7 +27,7 @@ import sys
 from multiprocessing import Process
 
 # add path to the Cyphon project folder so Cyphon packages can be found
-CYPHON_PATH = os.path.dirname(os.path.abspath(__file__))
+CYPHON_PATH = os.path.dirname(os.path.dirname(os.path.abspath(__file__)))
 sys.path.append(CYPHON_PATH)
 sys.path.append(os.path.dirname(os.path.dirname(CYPHON_PATH)))
 
@@ -43,23 +43,78 @@ import django
 django.setup()
 
 # third party
-from django.core.exceptions import ObjectDoesNotExist
 import pika
 
 # local
+from cyphon.documents import DocumentObj
 from cyphon.transaction import close_connection, close_old_connections
+from monitors.models import Monitor
+from sifter.datasifter.datachutes.models import DataChute
 from sifter.logsifter.logchutes.models import LogChute
-from sifter.logsifter.logmungers.models import LogMunger
+from watchdogs.models import Watchdog
 
 LOGGER = logging.getLogger('receiver')
 
-LOGSIFTER = settings.LOGSIFTER
 BROKER = settings.RABBITMQ
 
 
-def consume_logs():
+def _create_doc_obj(body):
+    """Turn a message str into a DocumentObj."""
+    data = json.loads(body)
+    doc_id = data.get('@uuid')
+    collection = data.get('collection')
+    return DocumentObj(data=data, doc_id=doc_id, collection=collection)
+
+
+@close_connection
+def process_msg(channel, method, properties, body):
+    """Process a message.
+
+    Callback function for a queue consumer.
+
+    Parameters
+    ----------
+    channel : pika.Channel
+
+    method : pika.spec.Basic.Return
+
+    properties : pika.spec.BasicProperties
+
+    body : str, unicode, or bytes (python 3.x)
+
+    routing_key : str
+        Indicates the type of consumer to use. Options are 'datachutes',
+        'logchutes', 'monitors', 'watchdogs'.
+
     """
-    Creates a queue consumer for RabbitMQ.
+    consumers = {
+        'datachutes': DataChute.objects.process,
+        'logchutes': LogChute.objects.process,
+        'monitors': Monitor.objects.process,
+        'watchdogs': Watchdog.objects.process,
+    }
+
+    try:
+        if not isinstance(body, str):
+            body = body.decode('utf-8')
+
+        doc_obj = _create_doc_obj(body)
+        consumer_func = consumers[method.routing_key]
+        consumer_func(doc_obj)
+
+    except Exception as error:
+        LOGGER.exception('An error occurred while processing the message '
+                         '\'%s\':\n  %s', body, error)
+
+
+def consume_queue(routing_key='watchdogs'):
+    """Create a queue consumer for RabbitMQ.
+
+    Parameters
+    ----------
+    routing_key : str
+        Options are 'DATACHUTES', 'LOGCHUTES', 'MONITORS', 'WATCHDOGS'.
+
     """
     try:
         credentials = pika.PlainCredentials(username=BROKER['USERNAME'],
@@ -76,9 +131,9 @@ def consume_logs():
         channel = conn.channel()
         exchange = BROKER['EXCHANGE']
         exchange_type = BROKER['EXCHANGE_TYPE']
-        routing_key = BROKER['ROUTING_KEY']
         durable = BROKER['DURABLE']
-        queue_name = BROKER['QUEUE_NAME']
+
+        queue_name = routing_key
 
         channel.exchange_declare(exchange=exchange,
                                  type=exchange_type,
@@ -90,8 +145,8 @@ def consume_logs():
                            queue=queue_name,
                            routing_key=routing_key)
 
-        LOGGER.info('Waiting for logs')
-        # print(' [*] Waiting for logs. To exit press CTRL+C')
+        LOGGER.info('Waiting for messages')
+        # print(' [*] Waiting for messages. To exit press CTRL+C')
 
         channel.basic_consume(process_msg,
                               queue=queue_name,
@@ -100,74 +155,39 @@ def consume_logs():
         channel.start_consuming()
 
     except Exception as error:
-        LOGGER.exception('An error occurred while consuming logs:\n  %s', error)
-
-
-@close_connection
-def process_msg(channel, method, properties, body):
-    """
-    Callback function for a queue consumer.
-
-    Parameters:
-        channel: pika.Channel
-        method: pika.spec.Basic.Return
-        properties: pika.spec.BasicProperties
-        body: str, unicode, or bytes (python 3.x)
-
-    """
-    try:
-        if not isinstance(body, str):
-            body = body.decode('utf-8')
-
-        doc = json.loads(body)
-        data = doc.get('message')
-        doc_id = doc.get('@uuid')
-        collection = doc.get('collection')
-        saved = False
-        enabled_chutes = LogChute.objects.find_enabled()
-
-        LOGGER.info('Processing message %s in %s', doc_id, collection)
-
-        for chute in enabled_chutes:
-            result = chute.process(data, doc_id, collection)
-            if result:
-                saved = True
-
-        if not saved and LOGSIFTER['DEFAULT_LOG_CHUTE_ENABLED']:
-            default_munger = get_default_munger()
-            if default_munger:
-                default_munger.process(data, doc_id, collection)
-
-    except Exception as error:
-        LOGGER.exception('An error occurred while processing the message:\n  %s',
-                         body)
-
-
-def get_default_munger():
-    """
-    Returns the default LogMunger specified in the site configuration.
-    """
-    munger_name = LOGSIFTER['DEFAULT_LOG_MUNGER']
-    try:
-        return LogMunger.objects.get(name=munger_name)
-    except ObjectDoesNotExist:
-        LOGGER.error('Default LogMunger "%s" is not configured.', munger_name)
+        LOGGER.exception('An error occurred while consuming messages:\n  %s',
+                         error)
 
 
 @close_old_connections
-def create_consumers(num):
-    """
+def create_consumers(routing_key, num):
+    """Create one or more queue consumers.
+
+    Parameters
+    ----------
+    routing_key : str
+        Indicates how the messages will be processed. Options are
+        'datachutes', 'logchutes', 'monitors', 'watchdogs'.
+
+    num : str
+        A string representation of an integer representing the number of
+        consumers to spawn.
 
     """
+    kwargs = {'routing_key': routing_key}
     for dummy_num in range(num):
-        process = Process(target=consume_logs)
+        process = Process(target=consume_queue, kwargs=kwargs)
         process.start()
 
 
 if __name__ == '__main__':
     try:
-        NUM = int(sys.argv[1])
+        _ROUTING_KEY = sys.argv[1]
+    except IndexError:
+        _ROUTING_KEY = 'watchdogs'
+    try:
+        _NUM = int(sys.argv[2])
     except (IndexError, ValueError):
-        NUM = 1
+        _NUM = 1
 
-    create_consumers(NUM)
+    create_consumers(_ROUTING_KEY, _NUM)