+++ cyphon/sifter/datasifter/datachutes/tests/test_models.py
@@ -20,12 +20,13 @@ Tests the DataChute class.
 
 # standard library
 import logging
-from unittest.mock import Mock
+from unittest.mock import call, Mock, patch
 
 # third party
 from django.test import TestCase
 
 # local
+from cyphon.documents import DocumentObj
 from sifter.datasifter.datachutes.models import DataChute
 from tests.fixture_manager import get_fixtures
 
@@ -39,8 +40,30 @@ class DataChuteTestCase(TestCase):
     def setUp(self):
         logging.disable(logging.ERROR)
 
+        # clear cached property
+        try:
+            del DataChute.objects._default_munger
+        except AttributeError:
+            pass
+
     def tearDown(self):
         logging.disable(logging.NOTSET)
+
+    @patch('sifter.datasifter.datachutes.models.DataChute.process')
+    def test_bulk_process(self, mock_process):
+        """
+        Tests the bulk_process method for a chute.
+        """
+        data_1 = {}
+        data_2 = {}
+        mock_doc_1 = Mock()
+        mock_doc_2 = Mock()
+        with patch('sifter.datasifter.datachutes.models.DocumentObj',
+                   side_effect=[[mock_doc_1], [mock_doc_2]]) as mock_doc:
+            datachute = DataChute.objects.get(pk=1)
+            datachute.bulk_process([data_1, data_2])
+            mock_doc.has_calls([call(data_1), call(data_2)])
+            mock_process.has_calls([call(mock_doc_1), call(mock_doc_2)])
 
     def test_process_match(self):
         """
@@ -49,14 +72,14 @@ class DataChuteTestCase(TestCase):
         mock_doc_id = 1
 
         data = {'id': 123, 'subject': 'This is a Critical Alert'}
+        doc_obj = DocumentObj(data=data)
 
         datachute = DataChute.objects.get(pk=3)
         datachute.munger.process = Mock(return_value=mock_doc_id)
 
-        doc_id = datachute.process(data)
+        doc_id = datachute.process(doc_obj)
 
-        datachute.munger.process.assert_called_once_with(data, None, None,
-                                                         'twitter')
+        datachute.munger.process.assert_called_once_with(doc_obj)
         self.assertEqual(doc_id, mock_doc_id)
 
     def test_process_nonmatch(self):
@@ -64,11 +87,12 @@ class DataChuteTestCase(TestCase):
         Tests the process method for a nonmatching data dictionary.
         """
         data = {'id': 123, 'Subject': 'This is an Urgent Alert'}
+        doc_obj = DocumentObj(data=data)
 
         datachute = DataChute.objects.get(pk=3)
         datachute.munger.process = Mock(return_value=None)
 
-        doc_id = datachute.process(data)
+        doc_id = datachute.process(doc_obj)
 
         self.assertEqual(doc_id, None)
 
@@ -79,13 +103,12 @@ class DataChuteTestCase(TestCase):
         mock_doc_id = 1
 
         data = {'id': 123, 'Subject': 'This is an Urgent Alert'}
+        doc_obj = DocumentObj(data=data)
 
         datachute = DataChute.objects.get(pk=4)
         datachute.munger.process = Mock(return_value=mock_doc_id)
 
-        doc_id = datachute.process(data)
+        doc_id = datachute.process(doc_obj)
 
-        datachute.munger.process.assert_called_once_with(data, None, None,
-                                                         'twitter')
+        datachute.munger.process.assert_called_once_with(doc_obj)
         self.assertEqual(doc_id, mock_doc_id)
-